modified the structure
